third_party/blink/web_tests/http/tests/sendbeacon/beacon-detached-no-crash-expected.txt

Accessing Navigator sendBeacon methods on a closed window.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS w.navigator.sendBeacon() threw exception TypeError: Failed to execute 'sendBeacon' on 'Navigator': 1 argument required, but only 0 present..
PASS w.navigator.sendBeacon('resources/blank.txt', 'detached-should-return false') is false
PASS successfullyParsed is true

TEST COMPLETE
